requirements.txt
alabaster==0.7.3
alembic==0.9.1
altgraph==0.10.2
appnope==0.1.0
args==0.1.0
astroid==1.3.6
audioread==2.1.5
autoenv==1.0.0
Babel==1.3
backports.ssl-match-hostname==3.4.0.2
bcrypt==3.1.3
bdist-mpkg==0.5.0
beautifulsoup4==4.3.2
blessings==1.6
blinker==1.3
bonjour-py==0.3
bpython==0.14.2
bs4==0.0.1
certifi==2015.9.6.2
cffi==1.10.0
click==6.7
clint==0.4.1
commando==0.3.4
configparser==3.3.0.post2
coverage==4.0.3
curtsies==0.1.19
decorator==3.4.0
Django==1.8.5
django-admin-bootstrapped==2.5.7
django-bootstrap-form==3.2
django-bootstrap-forms==0.1
django-bower==5.0.2
django-cors-headers==1.0.0
django-nose==1.3
djangorestframework==3.1.1
docopt==0.4.0
docutils==0.12
ecdsa==0.13
emoji==0.4.5
enum34==1.1.6
Fabric==1.10.2
feedparser==5.1.3
Flask==0.12.1
Flask-API==0.7.1
Flask-Babel==0.9
Flask-Bcrypt==0.7.1
Flask-Login==0.3.0
Flask-Mail==0.9.0
Flask-Migrate==2.0.3
Flask-OpenID==1.2.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.10.2
flipflop==1.0
fswrap==0.1.2
funcsigs==1.0.2
functools32==3.2.3.post2
fuzzywuzzy==0.7.0
gnureadline==6.3.3
googlemaps==2.2
greenlet==0.4.9
guess-language==0.2
gunicorn==19.1.1
httplib2==0.9
httpretty==0.8.8
hyde==0.8.8
ipykernel==4.0.3
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.9.6
joblib==0.11
jsonschema==2.5.1
jupyter-client==4.0.0
jupyter-core==4.0.6
librosa==0.5.1
line-profiler==2.0
llvmlite==0.21.0
logging==0.4.9.6
logilab-common==0.63.2
macholib==1.5.1
Mako==1.0.6
mandrill==1.0.57
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==1.3.1
migrate==0.3.7
mistune==0.7.1
mock==1.0.1
modulegraph==0.10.4
mps==0.20.16
mutagen==1.27
nbconvert==4.0.0
nbformat==4.0.0
ndg-httpsclient==0.4.0
nltk==3.0.1
nose==1.3.4
notebook==4.0.5
numba==0.36.2
numpy==1.8.0rc1
oauth2==1.5.211
oauthlib==0.7.2
pandas==0.15.2
paramiko==1.15.2
path.py==8.1.1
pbr==1.8.1
pdfminer==20140328
pep8==1.5.7
pexpect==3.3
piazza-api==0.5.1
pickleDB==0.4
pickleshare==0.5
Pillow==2.7.0
praw==2.1.20
psycopg2==2.7.1
ptyprocess==0.5
py2app==0.7.3
pyasn1==0.1.8
PyAudio==0.2.8
pybing==0.12
pycorenlp==0.3.0
pycparser==2.17
pycrypto==2.6.1
Pygments==2.0.2
PyJWT==1.4.2
pylint==1.4.3
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-coveralls==2.9.1
python-dateutil==1.5
python-editor==1.0.3
python-gmaps==0.2.1
python-openid==2.2.5
python-tesseract==0.9.1
python-twitter==2.2
PyTumblr==0.0.6
pytz==2013.7
PyYAML==3.12
pyzmq==14.7.0
quora==0.1.22
requests==2.13.0
requests-oauthlib==0.4.2
resampy==0.2.0
scikit-learn==0.17
scipy==0.13.0b1
simplegeneric==0.8.1
simplejson==3.6.5
singledispatch==3.4.0.3
six==1.4.1
sklearn==0.0
Skype4Py==1.0.35
smartypants==1.7.1
snowballstemmer==1.2.0
speaklater==1.3
SpeechRecognition==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
spotipy==2.4.4
SQLAlchemy==1.1.9
stanford-corenlp-pywrapper==0.1.0
stevedore==1.1.0
Tempita==0.5.2
terminado==0.5
tornado==4.2.1
traitlets==4.0.0
twilio==3.8.0
typogrify==2.0.0
Unidecode==0.4.17
update-checker==0.11
vboxapi==1.0
virtualenv==15.1.0
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
Werkzeug==0.12.1
Whoosh==2.6.0
wikipedia==1.4.0
WTForms==2.0.1
xattr==0.6.4
xlrd==0.9.3
XlsxWriter==0.6.6
youtube-dl==2015.6.25
zope.interface==4.1.1
